FIx replication with encrypted WAL
walsender may read the beginning of the page directly from the buffer
and bypass tdeheap_xlog_seg_read. In that case, the reader of the page
reminder wouldn't know that it is encrypted and pass it to the
walreciever as it is (encrypted). To fix this, we save the last
encrypted header for tdeheap_xlog_seg_read.